FAM generation finished

+++ src/main/kotlin/models/FAMCell.kt
@@ -2,5 +2,5 @@ package models
 
 data class FAMCell(
         val antecedents: IntArray,
-        val output: Int
+        val output: Long
 )
